CORE-16319 - move statemanager api to an api folder to fix osgi issue. - remove some unnecessary dependencies
