+++ components/state-manager/src/main/java/net/corda/statemanager/service/package-info.java
@@ -1,4 +1,4 @@
 @Export
-package net.corda.statemanager;
+package net.corda.statemanager.service;
 
 import org.osgi.annotation.bundle.Export;
