+++ libs/state-manager/state-manager-api/src/main/java/net/corda/libs/statemanager/api/package-info.java
@@ -1,4 +1,4 @@
 @Export
-package net.corda.libs.statemanager;
+package net.corda.libs.statemanager.api;
 
 import org.osgi.annotation.bundle.Export;
