+++ libs/state-manager/state-manager-api/src/main/kotlin/net/corda/libs/statemanager/api/State.kt
@@ -1,4 +1,4 @@
-package net.corda.libs.statemanager
+package net.corda.libs.statemanager.api
 
 /**
  * A state managed via the state manager.
